Update Setup Command :building_construction:
Update the setup command to switch to new tabs _and_ run commands in
them. The previous technique was failing to gain focus on each new tab,
and would run subsequent commands in the original tab. Using `xdotool`
and the `key` command for a new tab also forces focus in the new tab.