Update Setup Command to Change Dir :new:
Update the daily setup command to also change dirs to the current
project. This introduces a new callback to set the current project that
is most frequently worked on, and then to have each new tab change to
that directory. The original setup command was only using the directory
from which the command was run, even if changed for the first tab.