Add Git Branch Name to Bash Prompt 📛
It is important to be aware of which git branch one is focused on. Its
also easy to lose track of which branch is currently active when working
on some deep problems or yak shaving.

This change adds the `__git_ps1` function to the bash prompt, which will
add the branch name to the prompt if available.

For more information, see [Ask Ubuntu regarding Git Prompt Script](https://askubuntu.com/questions/1521308/where-is-git-prompt-sh-located-in-ubuntu-24-04).